requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.1.1
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
requests==2.28.2
requests-html==0.10.0
six==1.16.0
soupsieve==2.4
tqdm==4.64.1
urllib3==1.26.14
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
